implement ingress and ServiceSpec.Ports[].Endpoint
This commit allows specifying a root level "ingress" field which
allows specifying an IngressSpec resource, plus a Name field for
the name of the ingress.

This also adds a _shortcut_ "endpoint" in
"services[].ports[].endpoint" which is essentially an extension of
"ServiceSpec.Ports[]" which allows specifying the Ingress' Host
and Path in the form "ingress_host/ingress_path", and an Ingress
resource is automatically generated based on the Service Port.

The name of the resulting Ingress is
"ServiceName-ServicePort.Port"

Fixes #22