Auto populating port names, if not specified
Before this commit, if a definition is passed with multiple ports
in one service, then in the generated Kubernetes output, the port
names are not populated, which when deployed to a Kubernetes
cluster, throws an error.

This commit fixes this, such that if multiple ports for one
service are specified, then if the port names are not specified,
they are auto generated and populated, in the format -
"service_name-service_port"

Note that, the auto generation and population only happens in
case of multiple ports, and in the case when the port names
are not already specified.

e.g. - for the input -

name: test
containers:
 - image: nginx
services:
- name: nginx
  ports:
  - port: 8080
  - port: 8081

the resulting output will contain Kubernetes service as -

apiVersion: v1
kind: Service
metadata:
  creationTimestamp: null
  labels:
    app: test
  name: nginx
spec:
  ports:
  - name: nginx-8080
    port: 8080
    targetPort: 0
  - name: nginx-8081
    port: 8081
    targetPort: 0
  selector:
    app: test

This commit also adds tests for this behavior.

Fix #38