Refactor modified PodSpec to PodSpecMod
The Kubernetes specs which we modify or add shortcuts to, are
in the format of <spec>Mod struct, example ServicePortMod,
IngressSpecMod, etc. This was not done for PodSpec in the App
struct till now.

This commit refactors the modified PodSpec out of the App struct
into PodSpecMod struct, and refactors tests.

Fixes #112