feat(spec): add support for extraResources
With this root level field called `extraResources`, which is a list
of file names of Kubernetes artifacts that can be fed to Kubernetes
directly without kedge having to do any processing.

Anything that is not supported in kedge can be provided using this
field. Following is the example snippet of the app file that is
using `extraResources` field to deploy a `secret.yaml` and a
`cron-job.yaml` with the rest of kedge file. Here `secret.yaml` &
`cron-jobs.yaml` reside in the same directory.

e.g.

```
...
extraResources:
- secret.yaml
- cron-jobs.yaml
```