move files from package encoding to package spec
This commit is purely a refactor. Some files from pkg/encoding
have been moved to pkg/spec. This is the first part of a bigger
effort to consolidate all the controller specific functions and
methods under one package, i.e. package spec for now, so that we
can start defining new controllers on a codebase that's more
suited for extensibility.