move files from package kubernetes to package spec
This commit is purely a refactor. All files from
pkg/transform/kubernetes have been moved to pkg/spec. This is
the second part of a bigger effort to consolidate all the
controller specific functions and methods under one package,
i.e. package spec for now, so that we can start defining new
controllers on a codebase that's more suited for extensibility.