add shortcut for portMappings
This commit adds a shortcut to ServiceSpec called portMappings.
This lets us set the port, targetPort and the protocol for a
service in a single line. This is parsed and converted to a
Kubernetes ServicePort object.

portMappings is an array of `port:targetPort/protocol`
definitions, so the syntax looks like -

portMappings:
- <port:targetPort/protocol>
- <port:targetPort/protocol>

Also, tests are added to test the added behavior.

The logic of auto populating names for the services has been moved
from fix.go to populators.go, because now the auto population
happens once the []ServicePort is populated from the both the
inputs, ServiceSpecMod.Ports and ServiceSpecMod.PortMappings

Fixes #216