merge ObjectMeta with k8s resource definitions
This commit merges ObjectMeta with at the rool level of Kubernetes
resource definitions wherever it's applicable.

This is done to extend the support for Kubernetes API further and
not ending up limiting the end user from defining certain fields
for their application which are in ObjectMeta.

For instance, before this commit, only `name` and `labels` were
supported, but now after this merge, all of the fields like
`annotations` or `deletionGracePeriodSeconds` which are present in
ObjectMeta can be defined at the root level.

Fixes #262