Script to display live output from e2e tests
Currently, when e2e tests are run, the output is displayed only
after the test run is complete. This is a known limitation of
`go tests`.

With this script, we run `make test-e2e` and
`kubectl get po --all-namespaces -w` in the background, and
terminate the process once the test run is complete, helping us
keep a watch on the individual tests while they are being run.