Add DeploymentConfig support
Adds DeploymentConfig support to Kedge by being able to specify a
controller like so:

```yaml
controller: deploymentconfig
```

Unfortunatley due to conflicting vendoring, a "hack" of openshift and
kubernetes types is added in order to prevent the full importation of
kubernetes/kubernetes which uses conflicting vendored files between
client-go, apimachinery and the main kubernetes/kubernetes directory.

Within the `/hack` directory is code copied and pasted from both
Kubernetes and OpenShift in order for us to use only what we need (the
structs).