Script to display live output from e2e tests: (#275)
Currently, when e2e tests are run, the output is displayed only
after the test run is complete. This is a known limitation of go tests.

With this simple script, when we run
 `go test -v github.com/kedgeproject/kedge/tests/e2e`, `make test-e2e`
and `kubectl get po --all-namespaces -w` are in the background, and
terminate the kubectl process once the test run is complete,
helping us keep a watch on the individual tests while they are being run.