add support for OpenShift Routes
This commit adds support for defining OpenShift routes in a Kedge
file.

Routes can be defined like the following at the root level of a
Kedge file

routes:
- name: route1
  host: route1-myproject.192.168.42.69.nip.io
  port:
    targetPort: 8080
  to:
    kind: Service
    name: httpd

This is implemented by essentially merging RouteSpec and ObjectMeta
and adding an array of the resulting struct at the root level under
the field "routes"

Adds docs, tests.

Fixes #238