add OpenShift Apps scheme for DeploymentConfig
This commit adds OpenShift's Apps scheme to support
DeploymentConfig.

Prior to this commit, TypeMeta for a DeploymentConfig had to
be hard coded, but now it's derived from the generated scheme
after addition of the Apps scheme.

The code in Transform() method in pkg/spec/deploymentconfig.go is
also refactored to leverage the scheme and be aligned with the
code in pkg/spec/deployment.go

Fixes #349