always populate service port name
Service port name was only populated when, there were more than one
ports in the service, but now the port names will be always be
populated if none given by the user.

When you create a service with just one service and expose it with a
OpenShift route, and later in time if you add one more port to the
exposed service, the OpenShift route will send traffic to all the
ports of service. To avoid that naming the port from the first time
is necessary.