Make small optimizations to the Dockerfile
This commit makes some changes to the Dockerfile so we do not have
to pull kedge from GitHub all over again since we are copying the
local code to inside the image anyway.

This should be fine since we are building the image from the make
command, so it's safe to assume that the end user has cloned kedge
already and running then running the `make image` command.